Add map based tables.
map[string]* tables have custom names.

If a map key type implements a String() method, it's tests will also
have custom names.

For other map key types, the type of the key, and a number is used.
This may not be the best idea. It will probably change after someone
uses it and has given feedback.